Layout is converted from client sided to server sided.
Also removed mentor page.
