--- src/app/admin/mentors/page.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import { Box } from '@mui/material';
-
-const MentorsPage: React.FC = () => {
-  return (
-    <Box>
-      <h1>Mentors</h1>
-    </Box>
-  );
-};
-
-export default MentorsPage;
